src/inventory/requirements.txt
flask==3.1.0
grpcio==1.68.1
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-instrumentation-flask==0.50b0
opentelemetry-instrumentation-grpc==0.50b0
opentelemetry-instrumentation-requests==0.50b0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-semantic-conventions==0.50b0
requests==2.32.3
